+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Changed
 
 - Updated readme.
-- Use base version for start and download commands if self-building in CI.
-- Remove unnecessary toolchain start on `update` command.
+- Use base version for `start` and `download` actions if self-building in CI.
+- Remove unnecessary toolchain start on `update` action.
 - Start correct version on `buildDragon`.
 
 ## [2.0.2] - 2019-30-11